+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Rework asserts in FSM library.
 - Rename state machine signals: `AM_EVT_HSM_` -> `AM_EVT_`, `AM_EVT_FSM_` -> `AM_EVT_`.
 - Set ticker task to the highest priority in workers example.
+- Merge queue library into event library.
 
 ### Added